\name{plot_clustered_data}
\alias{plot_clustered_data}
\title{
  Plot Clustered Data
}
\usage{
  \code{plot_clustered_data(X, model)}
}
\arguments{
  \item{X}{Data matrix.}

  \item{model}{A list containing the parameters of the Poisson mixture model.}
}
\description{
  Plots the clustered data based on a fitted Poisson mixture model.
}
