\name{predict}
\alias{predict}
\title{
  Predict Cluster Membership
}
\usage{
  \code{predict(model, X)}
}
\arguments{
  \item{model}{A list containing the parameters of the Poisson mixture model.}

  \item{X}{Data matrix.}
}
\description{
  Predicts the cluster membership of data points based on a fitted Poisson mixture model.
}
